@@ -1,5 +1,5 @@
 {
-    "name": "dgame/php-soap",
+    "name": "demvsystems/php-soap",
     "description": "php soap",
     "license": "MIT",
     "type": "library",
@@ -17,11 +17,12 @@
     ],
     "require": {
         "ext-json": "*",
-        "php": "^7.1",
-        "dgame/php-ensurance": "^1.0 || ^2.0"
+        "ext-dom": "*",
+        "php": "^8.1",
+        "dgame/php-ensurance": "^2.3"
     },
     "require-dev": {
-        "phpunit/phpunit": "^7.5"
+        "phpunit/phpunit": "^9"
     },
     "autoload": {
         "classmap": [
@@ -37,7 +38,7 @@
             "phpunit"
         ],
         "coverage": [
-            "phpunit --coverage-clover=coverage"
+            "XDEBUG_MODE=coverage phpunit --coverage-clover=coverage"
         ]
     }
 }
