@@ -21,5 +21,5 @@ interface ValuedInterface
     /**
      * @return mixed
      */
-    public function getValue();
+    public function getValue(): mixed;
 }
